Started breaking out useValidation functions
